+++ tes.sql
@@ -439,29 +439,13 @@ CREATE TABLE tes_join_info (
 
 
 
--- -- 翻打凭条
--- drop table if exists tes_receipt;
--- CREATE TABLE tes_receipt (
---   id BIGINT(21) not NULL AUTO_INCREMENT COMMENT '主键',
---   scene_id bigint(21) not null comment '场次id',
---   content varchar(11)  COLLATE utf8mb4_unicode_ci  not NULL COMMENT '内容信息，一个随机数字',
---   type INT(11) not NULL DEFAULT -1 COMMENT '管理员类型， 1 部门， 2 机构， 3 岗位',
---   is_del INT(1) not NULL DEFAULT '-1' COMMENT '删除标志',
---   update_time DATETIME  DEFAULT CURRENT_TIMESTAMP  COMMENT '更新时间',
---   update_user_id BIGINT(21)  not NULL DEFAULT '-1'  COMMENT '更新人id',
---   create_user_id BIGINT(21) not NULL DEFAULT '-1'  COMMENT '创建人id',
---   create_time DATETIME DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
---   PRIMARY KEY(id)
--- ) ENGINE=INnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci COMMENT='参考人员';
-
-
 
 -- 凭条信息
 DROP TABLE IF EXISTS tes_receipt;
 CREATE TABLE tes_receipt(
   id BIGINT(21) not NULL AUTO_INCREMENT COMMENT '主键',
   no int(11) not null default -1 comment '序号',
-  number int(11) not null default 0 comment '数字',
+  number bigint(21) not null default 0 comment '数字',
   scene_id bigint(21) not null default -1 comment '场次id',
   status INT(2) not NULL DEFAULT '-1' COMMENT '1 可用',
   is_del INT(1) not NULL DEFAULT '-1' COMMENT '删除标志',
@@ -479,8 +463,8 @@ create table tes_receipt_answer(
   id BIGINT(21) not NULL AUTO_INCREMENT COMMENT '主键',
   receipt_id bigint(21) not null default -1 comment '凭条页的id',
   no int(11) not null default -1 comment '序号',
-  number int(11) not null default 0 comment '数字',
-  answer int(11) not null default -1 comment '答案',
+  number bigint(21) not null default 0 comment '数字',
+  answer bigint(21) not null default -1 comment '答案',
   scene_id bigint(21) not null default -1 comment '场次id',
   user_id bigint(21) not null default -1 comment '答题人id',
   status INT(2) not NULL DEFAULT '-1' COMMENT '1 可用',
